+++ src/components/sidebar/sidebar.style.js
@@ -1,6 +1,5 @@
 import styled, { css } from 'styled-components';
 import baseTheme from '../../style/themes/base';
-import sidebarClassicStyle from './sidebar-classic.style';
 import StyledIconButton from '../icon-button/icon-button.style';
 
 const sidebarSizes = {
@@ -37,8 +36,6 @@ const SidebarStyle = styled.div`
     right: 25px;
     top: 25px;
   }
-
-  ${sidebarClassicStyle}
 `;
 
 SidebarStyle.defaultProps = {